# Faster serde