--- a/benches/lib_benches.rs
+++ b/benches/lib_benches.rs
@@ -2,9 +2,9 @@
 
 extern crate test;
 
+use bincode::Options as _;
 use core::str::FromStr;
 use rust_decimal::Decimal;
-use bincode::Options as _;
 
 macro_rules! bench_decimal_op {
     ($name:ident, $op:tt, $y:expr) => {