--- a/src/serde_types.rs
+++ b/src/serde_types.rs
@@ -174,7 +174,7 @@ impl serde::Serialize for Decimal {
     where
         S: serde::Serializer,
     {
-        serializer.serialize_str(&self.to_string())
+        serializer.serialize_str(self.to_array_str().as_ref())
     }
 }