I had a use case where my bottleneck was due to decimal deserialization via bincode, so I though I would do some low hanging fruit optimizations.

Feedback would be much appreciated.

## Overview
* Adds arrrayvec as a dependency since it will replace the heap allocated vectors and strings
* Removes heap allocation for both serialization & deserialization with the `serde-str` feature enable. This is done my removing heap allocation for the from_str impl as well as using a to_array_str impl instead of using to_string.
* Adds a from_str_radix specialization when radix is 10, since its the most common case (~25% performance improvement)
* Improve the to_string algorithm by making it 2 pass instead of 3 pass

## Outstanding issues
* Need to add tests for the `to_array_str` method
* Im not sure the constants I used for my arrays are the right ones. Basically we wanna use the highest possible size necessary otherwise it will panic when pushing a sufficiently long decimal.

## Performance improvements according to the micro benchmarks:
```
before:
test deserialize_bincode     ... bench:       6,544 ns/iter (+/- 3,308)
test serialize_bincode       ... bench:      15,397 ns/iter (+/- 7,748)
test decimal_to_string       ... bench:       7,288 ns/iter (+/- 77)
test decimal_from_str        ... bench:       5,684 ns/iter (+/- 180)

after:
test deserialize_bincode     ... bench:       2,884 ns/iter (+/- 1,448)
test serialize_bincode       ... bench:       5,248 ns/iter (+/- 57)
test decimal_to_string       ... bench:       4,203 ns/iter (+/- 48)
test decimal_from_str        ... bench:       1,592 ns/iter (+/- 810)
```
I also tested against my real life application and the performance improvements where similar. I got from about 500k reads/s to 1.3M reads/s. I did not test the serialization path however.