@jean-airoldie 2020-12-29
Also just a general word of advise, when using a `Vec` I would recommend to always try to `Vec::with_capacity` with a over approximated capacity of what you would need. Otherwise your vec might allocate multiple times, which is expensive.

Also, you might want to checkout smallvec which is an hybrid between a vec and a arrayvec where you only allocate if you exceed the capacity. I would say its always preferable to using a vec in any performance sensitive setting, unless your allocating huge amounts of things.