@paupino 2020-12-30
Awesome - great timing too, as I'm starting to comb through some optimizations over the next few days using a similar approach. I'll pull it once it's passing and take a deeper look into the changes. 

Btw, I'm using `cargo make` now to try to automate some of the common functions/feature combinations in the project - you can run `cargo make test-all` to run various combinations to track down where it's failing (looks like pg parsing has some rounding differences?).