@jean-airoldie 2020-12-30
Looks like testing for postgres locally is gonna be a pain since you need to have a db running or something.