@paupino 2020-12-30
Since I'm in there already, I can run those tests for you. Let me do some digging to see what's going on.