@paupino 2020-12-30
Looks like you found it! I think we should add one more test to `it_can_parse_highly_significant_numbers()`, namely:

```rust
("1.234567890123456789012345678949999", "1.2345678901234567890123456789"),
```

This captures what was happening in pg.