@jean-airoldie 2020-12-30
Alright so i think I got it. So looks like the midpoint calculation were always wrong. I'll make it so the tests pass but I'll put some FIXME comments.