@paupino 2020-12-30
Hmm, I'm not sure that's the fix, but agree that something is a bit weird. My suspicion is that it's actually the fixed size array that is interfering with `coeff.len()` calcs. Interestingly, you can get them passing by changing the max precision from `29` to `30` however I want to understand what is triggering the issue if possible. I'll run the code a bit more: I can see a few areas for optimization so will take a stab at cleaning that up while I'm there (e.g. `a-Z` parsing isn't required for radix 10)