@jean-airoldie 2020-12-30
I got the same error from fixing the midpoint on the master branch directly (which doesn't use arrays) so I'm relatively sure that its the cause.