@jean-airoldie 2020-12-30
As for the the fix size array, it should behave like the vector expect it doesn't have all the methods that vec has and should panic if the capacity overflows. So if it was a issue caused by the arrays, they would panic.