@jean-airoldie 2020-12-30
Check this PR https://github.com/paupino/rust-decimal/pull/299 that shows that the issue is present on master.