@paupino 2020-12-30
Oh, I see what you mean. Cool, in that case it's not the midpoint stuff that needs changing (i.e. to 10 etc) - it's the estimated max size which is wrong. For `10` this should be 28. I'll need to recalculate these to just double check it all.