@paupino 2020-12-30
Cool; after playing around with it I can see some clear improvements. Made some small refactors when I was reviewing so did a ff-merge to get the PR in. Thank you for doing this - two clear wins in my opinion! (midpoint logic issue + speed improvements)